backend/pyproject.toml
[project]
name = "pdf-chinese-font-fixer-backend"
version = "1.0.0"
description = "Backend API for AI-powered PDF Chinese font repair"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "email-validator>=2.2.0",
    "uvicorn[standard]>=0.34.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.18",
    "google-auth>=2.37.0",
    "google-generativeai>=0.8.0",
    "httpx>=0.28.0",
    "aiofiles>=24.1.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
